[enhancement] better search
so far the search always started from the top of the document. So furhter hits could not be found.

With this change search will start searching from the current cursor position.

Also with short-cut F3 search will be continued with current search string.

However search will not wrap at the end of the document. One still has to move the cursor to the start of the document to continue search from the beginning.