Feature continue search (#5)
So far the search always started from the top of the document. So further hits could not be found.

With this change the search will start from the current cursor position.

Also with shortcut F3 search will can be continued to the next hit.

However search will not wrap at the end of the document. One still has to move the cursor to the start of the document to continue search from the beginning.